requirements.txt
python-telegram-bot==21.9
python-dotenv==1.0.0
requests==2.31.0
authentik-client==2025.10.3
qrcode==8.0
pillow==11.0.0
